// src/utils/query/useColorsRouteQuery/useColorsRouteQuery.types.ts
export enum EQueriesColors {
  COLOR1 = "501599",
  COLOR2 = "87ABB7",
  COLOR3 = "B0BFCF",
  COLOR4 = "450F71",
  COLOR5 = "34A0B5",
  COLOR6 = "FA1A69",
  COLOR7 = "5A9870",
  COLOR8 = "9BB471",
  COLOR9 = "246704",
  COLOR10 = "3C872E",
  COLOR11 = "91B372",
  COLOR12 = "E5AA6D",
  COLOR13 = "F06BB8",
  COLOR14 = "D52BD3",
  COLOR15 = "86811B",
  COLOR16 = "3D9A79",
  COLOR17 = "659722",
  COLOR18 = "8EC8B5",
  COLOR19 = "E8118F",
  COLOR20 = "216DC1",
  COLOR21 = "9CC171",
  COLOR22 = "F7EDAB",
  COLOR23 = "000000",
}